+++ glean-core/android/src/main/java/mozilla/telemetry/glean/scheduler/GleanLifecycleObserver.kt
@@ -28,15 +28,6 @@ internal class GleanLifecycleObserver : LifecycleEventObserver {
                 // on foreground.
                 GleanBaseline.duration.stop()
                 Glean.handleBackgroundEvent()
-
-                // Clear the "dirty flag" as the last thing when going to background.
-                // If the application is not being force-closed, we should still be
-                // alive and allowed to change this. If we're being force-closed and
-                // don't get to this point, next time Glean runs it will be detected.
-                @Suppress("EXPERIMENTAL_API_USAGE")
-                Dispatchers.API.launch {
-                    LibGleanFFI.INSTANCE.glean_set_dirty_flag(false.toByte())
-                }
             }
             Lifecycle.Event.ON_START -> {
                 // Updates the baseline.duration metric when entering the foreground.
@@ -50,13 +41,8 @@ internal class GleanLifecycleObserver : LifecycleEventObserver {
                 // Note that this is sending the length of the last foreground session
                 // because it belongs to the baseline ping and that ping is sent every
                 // time the app goes to background.
+                Glean.handleForegroundEvent()
                 GleanBaseline.duration.start()
-
-                // Set the "dirty flag" to `true`.
-                @Suppress("EXPERIMENTAL_API_USAGE")
-                Dispatchers.API.launch {
-                    LibGleanFFI.INSTANCE.glean_set_dirty_flag(true.toByte())
-                }
             }
             else -> {
                 // For other lifecycle events, do nothing