+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 [Full changelog](https://github.com/mozilla/glean/compare/v26.0.0...master)
 
 * General:
+  * The baseline ping is now sent when the application goes to foreground, rather than background. The only ping "reason" for the baseline ping is "foreground", and the "background" and "dirty-startup" are no longer used.
   * Glean will now detect when the upload enabled flag changes outside of the application, for example due to a change in a config file. This means that if upload is disabled while the application wasn't running (e.g. between the runs of a Python command using the Glean SDK), the database is correctly cleared and a deletion request ping is sent. See [#791](https://github.com/mozilla/glean/pull/791).
   * The `events` ping now includes a reason code: `startup`, `background` or `max_capacity`.
 * iOS: